refactor(ui): extract OwnPilotServerSection out of McpServersPage
OwnPilotServerSection (the OwnPilot-MCP-server status panel) is a
self-contained sub-component — no props, own state, no coupling to other
module symbols. Move it to its own file; McpServersPage.tsx drops from
1425 to 1089 lines. Now-unused icon/EmptyState/type imports pruned from
the page. Behaviour unchanged.

UI typecheck + lint + 358 tests + build green.
